datadog_agent_config_overlay_model

Module smoke_test_support

Types necessary to keep the configuration smoke tests working. Intended to be transitional until we can get Saluki/ADP to stop deserializing directly from Agent configuration. At that time the smoke tests will be rendered inert and can be removed.

Enums

ConfigurationStruct

A struct in the Saluki repository that deserializes Datadog Agent configuration.